app-el/requirements.txt
numpy==1.19.1
pandas==1.3.5
google-api-python-client==2.92.0
google-cloud-storage
openpyxl==3.0.0
